Add bug fix and pass tests

+++ src/matcha_ml/storage/azure_storage.py
@@ -125,6 +125,8 @@ class AzureStorage:
         container_client = self._get_container_client(container_name)
 
         for blob in container_client.list_blobs():
+            if "matcha.lock" in str(blob.name):
+                continue
             blob_client = container_client.get_blob_client(blob=str(blob.name))
             file_path = os.path.join(dest_folder_path, str(blob.name))
 
@@ -202,6 +204,9 @@ class AzureStorage:
 
         # Remove blobs that are not present in the local `src_folder_path``
         for blob in blob_set:
+            # Ensure that the lock file is not being prematurely removed from the remote bucket
+            if "matcha.lock" in blob:
+                continue
             container_client.delete_blob(blob)
 
     def _sync_local(self, dest_folder_path: str) -> None:
